+++ portal/src/app/components/protocol-query/ProtocolQuery.module.scss
@@ -100,19 +100,3 @@ $titlesColor: #1D2329;
 .required {
   color: $required;
 }
-
-.iterations_input_option {
-  margin-inline-end: 10px;
-  cursor: pointer;
-}
-
-.algorithms_input_option {
-  margin-inline-end: 10px;
-  margin-inline-start: 25px;
-  cursor: pointer;
-}
-
-.algorithms_input_option_title {
-  margin-inline-end: 10px;
-  cursor: pointer;
-}
