+++ portal/src/app/components/protocol-query/ProtocolQuery.tsx
@@ -1,6 +1,6 @@
 import { noop } from 'lodash';
 import React, { useCallback, useState } from 'react';
-import { OptionProps, Options, components } from 'react-select';
+import { Options } from 'react-select';
 import { ITestParams } from '../../shared/models/quantum.interface';
 import { Button, ButtonActionType, ButtonSize, ButtonStyleType } from '../../shared/components/att-button';
 import { AttSelect, AttSelectOption } from '../../shared/components/att-select';
@@ -9,7 +9,7 @@ import { PROTOCOL_QUERY_EN } from './translate/en';
 import { Spinner, SpinnerSize } from '../../shared/components/att-spinner';
 import { useGetAlgorithms, useGetIterations } from './hooks';
 import { handleAlgorithmsSelection } from './utils';
-import { algorithmSections } from './constants';
+import { AlgorithmsSelectorCustomOption, IterationsSelectorCustomOption } from '../../shared/components/selector-custom-option';
 
 export type SelectOptionType = AttSelectOption | Options<AttSelectOption> | null;
 type onTextChangedEvent = (e: React.ChangeEvent<HTMLInputElement>) => void;
@@ -29,25 +29,22 @@ export const ProtocolQuery: React.FC<ProtocolQueryProps> = (props: ProtocolQuery
   
   const [experimentName, setExperimentName] = useState('');
   const [algorithms, setAlgorithms] = useState<SelectOptionType>();
-  const [prevSelectedValues, setPrevSelectedValues] = useState<string[]>([]);
   const [iterationsCount, setIterationsCount] = useState<SelectOptionType>();
+  const [prevSelectedValues, setPrevSelectedValues] = useState<string[]>([]);
 
   const onSubmitHandler = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
     onRunClick({ experimentName, algorithms: algorithms as SelectOptionType, iterationsCount: iterationsCount as SelectOptionType });
   };
 
-  const onExperimentNameChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
+  const onExperimentNameChanged: onTextChangedEvent = (event: React.ChangeEvent<HTMLInputElement>) => {
     setExperimentName(event.target.value);
   };
 
   const onAlgorithmsChanged: OnSelectChanged = useCallback((options: SelectOptionType): void => {
-    let selectedAlgorithms: Options<AttSelectOption> = options as Options<AttSelectOption>;
-    
     const { newSelectedOptions, selectedValues } = 
-      handleAlgorithmsSelection(selectedAlgorithms, algorithmOptions, algosBySection, prevSelectedValues);
+      handleAlgorithmsSelection(options, algorithmOptions, algosBySection, prevSelectedValues);
 
-    // Update the selected algorithms and the previous selected values
     setAlgorithms(newSelectedOptions);
     setPrevSelectedValues(selectedValues);
   }, [algosBySection, algorithmOptions, prevSelectedValues]);
@@ -57,40 +54,6 @@ export const ProtocolQuery: React.FC<ProtocolQueryProps> = (props: ProtocolQuery
     setIterationsCount(selectedIterationNum);
   }, []);
   
-
-  const AlgorithmsCheckboxOption: React.FC<OptionProps> = (props: OptionProps) => {
-    const isSectionTitle = algorithmSections.includes((props.data as AttSelectOption).value);
-    const optionStyle = isSectionTitle ? styles.algorithms_input_option_title : styles.algorithms_input_option;
-
-    return (
-      <components.Option {...props}>
-        <div>
-          <input
-            type="checkbox"
-            className={optionStyle}
-            checked={props.isSelected}
-            onChange={() => onAlgorithmsChanged} />
-        </div>
-        <span>{props.label}</span>
-      </components.Option>
-    );
-  };
-
-  const IterationsCheckboxOption: React.FC<OptionProps> = (props: OptionProps) => {
-    return (
-      <components.Option {...props}>
-        <div>
-          <input
-            type="checkbox"
-            className={styles.iterations_input_option}
-            checked={props.isSelected}
-            onChange={() => onIterationsNumChanged} />
-        </div>
-        <span>{props.label}</span>
-      </components.Option>
-    );
-  };
-
   return (
     <div className={styles.protocol_query_wrapper}>
       <div>
@@ -126,7 +89,7 @@ export const ProtocolQuery: React.FC<ProtocolQueryProps> = (props: ProtocolQuery
                 hideSelectedOptions={false}
                 closeMenuOnSelect={false}
                 required
-                customComponent={{ Option: AlgorithmsCheckboxOption as React.FC }}
+                customComponent={{ Option: AlgorithmsSelectorCustomOption as React.FC }}
               />
           </div>
           <div className={styles.form_item}>
@@ -143,7 +106,7 @@ export const ProtocolQuery: React.FC<ProtocolQueryProps> = (props: ProtocolQuery
                 hideSelectedOptions={false}
                 closeMenuOnSelect={false}
                 required
-                customComponent={{ Option: IterationsCheckboxOption as React.FC }}
+                customComponent={{ Option: IterationsSelectorCustomOption as React.FC }}
               />
           </div>
           <div className={styles.form_item}>
