+++ portal/src/app/shared/components/selector-custom-option/SelectorCustomOption.module.scss
@@ -1,0 +1,17 @@
+@import "src/styles/variables-keys";
+
+.algorithms_input_option {
+  margin-inline-end: 10px;
+  margin-inline-start: 25px;
+  cursor: pointer;
+}
+
+.algorithms_input_option_title {
+  margin-inline-end: 10px;
+  cursor: pointer;
+}
+
+.iterations_input_option {
+  margin-inline-end: 10px;
+  cursor: pointer;
+}
