+++ portal/src/app/shared/components/selector-custom-option/index.ts
@@ -1,0 +1,1 @@
+export * from './SelectorCustomOption';
